quickhelp-backend/requirements.txt
fastapi
uvicorn[standard]
supabase==2.4.0
python-dotenv
pydantic[email]